work_test/requirements.txt
adrf==0.1.9
asgiref==3.8.1
Django==4.2.19
djangorestframework==3.15.2
flake8==5.0.4
flake8-docstrings==1.7.0
pep8-naming==0.13.3
psycopg2-binary==2.9.10
pytils==0.4.1
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-subtests==0.9.0
python-dotenv==1.0.1